k8s_automation_script/requirements.txt
kubernetes
PyYAML
click
tabulate